+++ src/main/java/zd/zero/waifu/motivator/plugin/alert/TextualWaifuNotification.kt
@@ -6,7 +6,7 @@ import com.intellij.openapi.project.Project
 class TextualWaifuNotification (
         motivationAsset: MotivationAsset,
         private val project: Project
-): BaseNotification(motivationAsset) {
+): BaseWaifuNotification(motivationAsset) {
     override fun createNotification(): Notification {
         val notification = buildNotification()
         notification.notify(project)
