+++ src/main/java/zd/zero/waifu/motivator/plugin/alert/VisualWaifuNotification.kt
@@ -6,11 +6,12 @@ import com.intellij.openapi.project.Project
 import com.intellij.openapi.ui.popup.Balloon
 import com.intellij.ui.BalloonLayoutData
 import zd.zero.waifu.motivator.plugin.onboarding.BalloonTools.fetchBalloonParameters
+import zd.zero.waifu.motivator.plugin.service.ApplicationService
 
 class VisualWaifuNotification(
-        motivationAsset: MotivationAsset,
-        val project: Project
-) : BaseNotification(motivationAsset) {
+    motivationAsset: MotivationAsset,
+    val project: Project
+) : BaseWaifuNotification(motivationAsset) {
 
     override fun createNotification(): Notification {
         val updateNotification = buildNotification()
@@ -21,8 +22,9 @@ class VisualWaifuNotification(
                 updateNotification,
                 true,
                 true,
-                    BalloonLayoutData.fullContent()
-            ) {}
+                BalloonLayoutData.fullContent(),
+                ApplicationService.instance
+            )
             balloon.show(notificationPosition, Balloon.Position.atLeft)
         } catch (e: Throwable) {
             updateNotification.notify(project)
