+++ src/test/java/zd/zero/waifu/motivator/plugin/alert/WaifuMotivationAssetTest.java
@@ -7,7 +7,7 @@ import static org.mockito.Mockito.*;
 
 public class WaifuMotivationAssetTest {
 
-    private final WaifuMotivation motivatorAlert = mock( WaifuMotivation.class );
+    private final WaifuMotivation motivatorAlert = mock( TextualMotivation.class );
 
     @Before
     public void setup() {
